+++ firefoxThemes/options/options.js
@@ -1,44 +1,47 @@
+/*DOM Elements*/
 const loadOnStartCheckbox = document.querySelector("#loadOnStart");
-
+const textSelectionCheckbox = document.querySelector("#textSelection");
+const scrollbarCheckbox = document.querySelector("#scrollbar");
+const root = document.querySelector(':root');
 /*Set color of options menu based on theme*/
 function setCss(chosenTheme) {
-  if (!chosenTheme) return
-  const { colors } = chosenTheme.definition
-  const styles = `
-body, html {
-  background-color: ${colors.headerColor};
-  color: ${colors.foregroundColor};
+  if (!chosenTheme) return;
+  const {colors} = chosenTheme.definition;
+  root.style.setProperty('--checkbox-color',colors.accentColor);
+  root.style.setProperty('--foreground-color',colors.foregroundColor);
+  root.style.setProperty('--checkbox-border-color',colors.foregroundColor+70);
+  root.style.setProperty('--header-color',colors.headerColor);
+  root.style.setProperty('--checkmark-color',colors.classNameColor);
 }
 
-* {
-  background-color: ${colors.baseBackground};
- color: ${colors.foregroundColor};
+function initContent() {
+  browser.storage.local.get(['loadOnStart', 'textSelection', 'scrollbar', 'waifuThemes', 'currentThemeId'])
+    .then((storage) => {
+      initCheckbox(loadOnStartCheckbox,!!storage.loadOnStart);
+      initCheckbox(textSelectionCheckbox,!!storage.textSelection);
+      initCheckbox(scrollbarCheckbox,!!storage.scrollbar);
+      setCss(storage.waifuThemes.themes[storage.currentThemeId])
+    });
 }
 
-input:checked {
-  background-color: ${colors.accentColor}44;
-}
-
-label[for="loadOnStart"]
-{
-  color:${colors.foregroundColor};
-}
-        `
-  const styleSheet = document.createElement("style");
-  styleSheet.innerText = styles;
-  document.head.appendChild(styleSheet)
-}
-
-browser.storage.local.get(['loadOnStart', 'waifuThemes', 'currentThemeId'])
-  .then((storage) => {
-    loadOnStartCheckbox.checked = !!storage.loadOnStart;
-    setCss(storage.waifuThemes.themes[storage.currentThemeId])
+const onChangeCheckEvents = async (e) => {
+  changeCheckboxState(e);
+  await browser.runtime.sendMessage({
+    optionName: e.target.id,
+    optionValue: e.target.className
   });
-
-const onLoadSwitchChange = async () => {
-  await browser.storage.local.set({
-    loadOnStart: loadOnStartCheckbox.checked,
-  })
 };
+/*Initialize checkbox state*/
+function initCheckbox(el,state){
+  el.className = state ? 'checked': '';
+}
+/*Changes checkbox state*/
+function changeCheckboxState(e){
+  const el = e.target;
+  el.className = el.className !== 'checked' ? "checked":'';
+}
 
-loadOnStartCheckbox.addEventListener('change', onLoadSwitchChange, true)
+initContent();
+loadOnStartCheckbox.addEventListener('click', onChangeCheckEvents, true);
+textSelectionCheckbox.addEventListener('click', onChangeCheckEvents, true);
+scrollbarCheckbox.addEventListener('click', onChangeCheckEvents, true);